requirements.txt
# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
numpy>=1.24.0

# MIDI and FL Studio Communication
mido>=1.3.0
python-rtmidi>=1.5.0
pyflp>=2.2.1

# Music Theory and Generation
mingus>=0.6.1
music21>=9.1.0

# Natural Language Processing
spacy>=3.7.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
